Film Table v2
Added movie links
