+++ frontend/src/components/styles/PopularPeople.css
